Updated home.html
A catch-22 is a paradoxical situation from which an individual cannot escape because of contradictory rules or limitations.[1] The term was coined by Joseph Heller, who used it in his 1961 novel Catch-22.

Catch-22s often result from rules, regulations, or procedures that an individual is subject to, but has no control over, because to fight the rule is to accept it. Another example is a situation in which someone is in need of something that can only be had by not being in need of it (e.g. the only way to qualify for a loan is to prove to the bank that you do not need a loan). One connotation of the term is that the creators of the "catch-22" situation have created arbitrary rules in order to justify and conceal their own abuse of power.